move virtual filesystem to the directory vfs
The virtual filesystem should not be part of the system call
implementation.
