+++ src/lib.rs
@@ -84,7 +84,6 @@ mod env;
 pub mod errno;
 mod executor;
 pub(crate) mod fd;
-#[cfg(feature = "fs")]
 pub(crate) mod fs;
 mod mm;
 mod scheduler;
@@ -283,7 +282,6 @@ extern "C" fn initd(_arg: usize) {
 
 	syscalls::init();
 	fd::init();
-	#[cfg(feature = "fs")]
 	fs::init();
 
 	// Get the application arguments and environment variables.
